Move mutations to own class
